Add Get for Company

+++ src/main/java/anagrafica/dto/address/AddressRequest.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class AddressRequest {
     private String address;
     private String cap;
-    private Integer zoneId;
+    private Long zoneId;
     private String pr;
 }
